+++ CHANGELOG.md
@@ -2,11 +2,20 @@
 
 ## [Unreleased]
 
+## 2024-11-14 - [0.1.1]
+
+### Fixes
+
+- Exit with code 1 when `trailing-whitespace` hook fixes.
+- Exit with code 2 when `trailing-whitespace` hook fails.
+- Print fixed path and line numbers in `trailing-whitespace` hook.
+
 ## 2024-11-13 - [0.1.0]
 
 ### New features
 
 - Add `trailing-whitespace` hook.
 
-[Unreleased]: https://github.com/pre-commit-bin/pre-commit-bin-hooks/compare/v0.1.0...main
+[Unreleased]: https://github.com/pre-commit-bin/pre-commit-bin-hooks/compare/v0.1.1...main
+[0.1.1]: https://github.com/pre-commit-bin/pre-commit-bin-hooks/compare/340da045...v0.1.1
 [0.1.0]: https://github.com/pre-commit-bin/pre-commit-bin-hooks/compare/77e4b9bd...340da045